Исправления модуляции
