+++ index.html
@@ -232,7 +232,7 @@
 
   <!-- Поделючение JS файла -->
   <script src="https://cdn.jsdelivr.net/npm/lodash@4.17.20/lodash.min.js"></script>
-  <script src="js/main.js" type="module"></script>
+  <script src="./js/main.js" type="module"></script>
 
 </body>
 </html>
