+++ js/consts.js
@@ -1,0 +1,1 @@
+// Оставлю пока пустым, в будущем буду загружать сюда переменные
